+++ src/gwt/src/org/rstudio/studio/client/workbench/commands/Commands.java
@@ -783,6 +783,7 @@ public abstract class
    // Chat satellite window
    public abstract AppCommand popOutChat();
    public abstract AppCommand returnChatToMain();
+   public abstract AppCommand uninstallPositAssistant();
    public abstract AppCommand checkForPositAssistantUpdates();
 
    // Internal